+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Set appID and orgID for notifications [#268](https://github.com/rokwire/groups-building-block/issues/268) 
 
+- Override update notifications preferences [#270](https://github.com/rokwire/groups-building-block/issues/270)
+
 ## [1.6.1] - 2022-10-12
 ### Changed
 - Finish the transition process of splitting the group record and membership list on two separate collections [#238](https://github.com/rokwire/groups-building-block/issues/238)